Remove github restriction (#28)
Summary:
This removes the github.com restriction for `vendor = false` git sources by employing a strategy similar to cargo for uniquely naming arbitrary git source urls. The url is normalized so it is consistently hashed, then combined with the last path component (for most hosted git providers, the name of the repo/project) + `-` + the hex hash of the normalized url.

Pull Request resolved: https://github.com/facebookincubator/reindeer/pull/28

Reviewed By: zertosh, shayne-fletcher

Differential Revision: D49193981

Pulled By: dtolnay

fbshipit-source-id: 69fac52eec04a49c9a4c67d2bd86ee95e9673d30