Fix subtarget_or_path for non-vendor cases
Summary:
`subtarget_or_path` is used for when you need to reach into a crates
sources. For vendored-mode, this is easy - it's just the path on disk.
For non-vendored-mode, it's complicated. This diff fixes the case where
the source is local, and now fails when the source is git (since buck2
doesn't support `sub_targets` on `git_fetch` unlike `http_archive`).

Reviewed By: diliop

Differential Revision: D50440981

fbshipit-source-id: 2c0bc5dc742473ab4dd95f2d1069bae3341f80af