Merge RootPackage, BuildscriptBinary, & BuildscriptGenrule rules. Merge PlatformRustCommon::env
Summary: When merging generated rules for each universe, handle BuildscriptBinary and BuildscriptGenrule rules, since these must be configured differently in different universes. Merge the RootPackage rule so that we can use the supertarget for no_std universe CI. Merge PlatformRustCommon::env so that the root package does not generate Buck dependencies (via the `$(location)` macro) on crates which are not included in the no_std universe.

Reviewed By: diliop

Differential Revision:
D52176256

Privacy Context Container: L1122763

fbshipit-source-id: c5032dca2095840a8c6dcd421186b3b61cd523d5