Do not require crates to be listed both in include_crates and features
Summary:
Specifying `include_crates` in the config for a universe means that in that universe, you wish to exclude every crate listed in Cargo.toml except for those explicitly listed in `include_crates`. This is useful for a `no_std` universe, since it is difficult to enumerate all of the crates and features in third-party/rust's Cargo.toml which somehow pull in libstd, and much easier to explicitly list the crates which don't.

When you wish to exclude some crate from the default universe (for instance, `panic-abort`), you can mark it as optional in Cargo.toml. To include the crate in some other universe which uses `include_crates`, you must do two things: you need to specify it in the `include_crates` list, and you must also enable the implicit root-crate feature generated by the optional dep (i.e., you must also list the dependency crate name in the universe config's `features` list).

This second step seems unintuitive and not useful. Instead, we can mark every crate listed in `include_crates` as non-optional, so that there is no need to manually enable the optional dependency.

Reviewed By: dtolnay

Differential Revision: D52635802

fbshipit-source-id: ccdaf09b560bfef1101acf3591f411b37adcdc11