Add ruff_python_parser
Summary:
This is a good error recovering Python parser. There are a few projects where we are thinking of using it - Pyre experiments, Autodeps work. None committed yet. Aim to experiment, and if it is of no use, rip out.

I've pointed at the upstream, but also made a clone in case it ever goes away. However, so it can be upgraded in future, pointing at the upstream seems the right thing.

Reviewed By: zertosh

Differential Revision: D62024492

fbshipit-source-id: 7aae8ab554db9906d76b2eae96d62d850f08b8de