Extract Subtarget-related logic to module
Summary: The new `insert_all` call in buckify.rs in D73601139 is not expressible using the `let mut insert = |&SubtargetOrPath| ...` closure that we're currently using for collecting needed subtargets, because it needs to use a different argument type (`&Option<Subtarget>` or `Option<&Subtarget>` or `&Subtarget`) than the existing calls (`&SubtargetOrPath`). Closures cannot be generic and 2 closures cannot both capture the same mutable hashmap that the existing closure is using, so I need the more elaborate API of `CollectSubtargets`.

Reviewed By: diliop

Differential Revision: D73601140

fbshipit-source-id: 63519d59c4469ca61cb115a7bc20161a6612e048