Supply accurate manifest dir for local path dependencies
Summary:
This is the correct way to replace the hack that was done in D73389843.

The hack used the sources from `cargo.rust_binary(name = "...-build-script-build", srcs = ...)` to emit a filegroup with a particular target name, which would then be looked up using `rule_exists` later by the `buildscript_run` rule.

The correct way directly informs `buildscript_run` about what directory to use as the manifest directory for the build script execution. This shouldn't be bolted on to `cargo.rust_binary`. In particular, the build script execution might need to look at other files from the crate which are not sources of the build script *compilation*.

```lang=starlark
buildscript_run(
    name = "with_buildscript-0.0.0-build-script-run",
    package_name = "with_buildscript",
    buildscript_rule = ":with_buildscript-0.0.0-build-script-build",
    local_manifest_dir = "sources/with_buildscript",  # <--
    version = "0.0.0",
)
```

Reviewed By: diliop

Differential Revision: D73625775

fbshipit-source-id: 1440893b1a664fc8eb4df2ac3e7cf99d5e6644cc