Delete support for limiting which targets rustc_flags apply to
Summary:
Cargo does not have any way to produce rustc_flags from a build script which apply to only the package's library and not the binary, or the binary and not library. So I don't think it's something we need to support in reindeer. If a buildscript produces rustc flags, both the library and binary will always get those flags.

In fbsource this affects exactly 1 package: `py-spy`.

```lang=diff
 third_party_rust_binary(
     name = "py-spy-0.4.0+custom1-py-spy-technicianted-py-spy",
     srcs = [...],
     crate = "py_spy",
     crate_root = "vendor/py-spy-0.4.0+custom1/src/main.rs",
     edition = "2021",
     env = {...},
     features = ["unwind"],
     licenses = ["vendor/py-spy-0.4.0+custom1/LICENSE"],
     platform = {...},
+    rustc_flags = ["@$(location :py-spy-0.4.0+custom1-build-script-run[rustc_flags])"],
     visibility = [],
     deps = [...],
 )
```

Reviewed By: cjlongoria

Differential Revision: D73675689

fbshipit-source-id: 22252d1d6978a2ee069e3f5a3358f4f5bf462a7f