Support CARGO_MANIFEST_LINKS cargo environment variable
Summary: Unlike for CARGO_PKG_DESCRIPTION and CARGO_PKG_REPOSITORY, Cargo does not set a default value of "" for CARGO_MANIFEST_LINKS if the manifest has no `links` in it. It is just unset. Also it is only set for the execution of the build script, not during the compilation of the build script, nor compilation of the library.

Reviewed By: diliop

Differential Revision: D73744513

fbshipit-source-id: 3ce4b63ef707bfb3bb6905ee1f548c33e818ed79