Delete support for fixup_templates
Summary:
We do not use this feature anywhere internally since 2 years ago, and it has become burdensome to maintain because it is implemented using a `Serialize` impl on Reindeer's `FixupConfigFile` data structure, which constrains how far the representation that is best for Reindeer's internal use can diverge from the raw TOML data structure written by users.

If we reintroduce a fixup templates feature later, I recommend that it should be done using a handwritten template TOML file that gets baked into Reindeer via an `include_str!`; not a `Serialize` impl. This would be a better user experience than the current implementation anyway because it will be able to use explanatory comments and better formatting than what the toml crate's serializer can produce.

Reviewed By: diliop

Differential Revision: D73823865

fbshipit-source-id: c9d5fa4e4af51c138eae7ab13bc2293b182c4823