Conflate gen_srcs and rustc_flags, set outdir and rustc flags for both
Summary:
Among the fixups in fbsource, I did not find any that deliberately want only the rustc flags from a build script or only the generated sources from a build script without the other. While there are lots of fixups that contain only `buildscript.gen_srcs` or only `buildscript.rustc_flags`, that's only because a particular build script is known not to produce the other (or was, at one point). Not because supplying both the buildscript-generated outdir and buildscript-generated rustc flags to the corresponding library or binary build would be incorrect.

This diff changes Reindeer to treat any `gen_srcs` fixup or `rustc_flags` fixup as both a `gen_srcs` and `rustc_flags` fixup in preparation for an upcoming consolidation of these into a single combined `buildscript.run` fixup.

If it turns out in the future that disabling the rustc flags or outdir from a particular build script is necessary, that should be accommodated using an opt-out such as:

```lang=toml
[buildscript.run]
rustc_flags = false
```

Reviewed By: diliop

Differential Revision: D73823868

fbshipit-source-id: aa1c846836714bcbebe1f25f3d474b8e9ba82adb