Prohibit platform-specific buildscript build configuration
Summary:
Consider this fixup:

```lang=toml
[buildscript.build]
env = {COMMON = "1"}

['cfg(windows)'.buildscript.build]
env = {PERPLAT = "1"}
```

It is nonsensical for Reindeer to generate a set of targets like this:

```lang=starlark
rust_binary(
    name = "example-0.0.0-build-script-build",
    ...
    platform = {
        "linux-arm64": dict(
            env = {
                "COMMON": "1",
            },
        ),
        "windows-msvc": dict(
            env = {
                "COMMON": "1",
                "PERPLAT": "1",
            },
        ),
        ...
    },
)

buildscript_run(
    name = "example-0.0.0-build-script-run",
    buildscript_rule = ":example-0.0.0-build-script-build",
)
```

because the build-script-build target is configured for the **execution platform** of the build-script-run target. It makes sense for build-script-run to have platform-specific environment and other attributes, because the `select` that this boils down to will correctly apply those. But build-script-build will not. That `select` will only be looking at whatever is the execution platform. For instance you could have build-script-run configured for Linux and Windows, but using Linux as the execution platform for both. A `'cfg(windows)'.buildscript.build` would not have the intended semantics in that case.

After execution modifiers (T197016652) are available it is possible that we will be able to do better by conveying enough information from build-script-run's target platform into build-script-run's execution platform that build-script-build can apply `select` meaningfully.

Reviewed By: cjlongoria

Differential Revision: D74027863

fbshipit-source-id: 6a94c2468d7558b6f10037972724ab719d348b6b