Split out link style for buildscript.build
Summary:
Previously, if a bin crate had a build script, and the fixup specified a `link_style`, that link style would apply to **both** the Rust binary and the build script build. There is no reason this should be the case. There is no expectation that a binary should want to use the same link style as its build script. In general they will have very different dependency graphs, with the build script using host tools like bindgen and libclang, while the binary links against target code.

This diff makes the build script link style independently configurable. We need this for libclang on macOS, such as in the fixup for the `libproc` and `coreaudio-sys` crates.

```lang=toml
link_style = "shared"  # applies to rust_binary only

[buildscript.build]
link_style = "shared"  # applies to build-script-build only
```

Reviewed By: cjlongoria

Differential Revision: D74046797

fbshipit-source-id: 20b59901b5670c607b1a69c939224121b965d46d