Import wasmtime crate (library, not CLI)
Summary: The Horizon Standalone Runtime project is experimenting with hosting wasmtime as a virtual machine for providing scripting support for the Metaverse (e.g., the Horizon OS Shell). This diff imports the wasmtime crate, with features enabled that provide the right starting point while balancing that with binary size. Other diffs in this stack start exercising some of the code and provide initial, rudimentary testing that the vendored version works; further adjustments may be necessary as we start using more of the API.

Reviewed By: capickett

Differential Revision: D73400656

fbshipit-source-id: 053b557e227446ad5498240fbf233df44ce3d893