Remove rustsec support from Cargo build
Summary:
As described in {D76385343}, there are issues with the most recent release of `rustsec` being incompatible with the most recent release of `cargo` due to mutually exclusive `gix` backends. We can easily support this in Buck-based builds as described in D76385343, but there is no way in Cargo-based builds.

This diff turns off RustSec in Reindeer's open-source Cargo-based build. We will continue using RustSec as usual in Buck-based builds.

Reviewed By: diliop

Differential Revision: D76385344

fbshipit-source-id: a9fa85e7bc64b752fcd48ba67db45b68970a3f90