Eliminate superfluous sorting of deserialized Cargo manifests
Summary:
These 3 data structures (`Metadata`, `Manifest`, `Node`) are deserialized from the output of `cargo metadata`.

Nothing in the way that Reindeer uses these data structures relies on the collections changed here being sorted, and the output from `cargo metadata` is not going to contain duplicate JSON list elements in any of these fields. There is no point to deserializing them as `BTreeSet` as opposed to `Vec`.

Switching to `Vec` makes way for changing the representation of some of the contents in a way that won't necessarily be sortable, notably for `PlatformExpr`.

Reviewed By: diliop

Differential Revision: D77062710

fbshipit-source-id: 88db878e894697fa2449737d2ebdd5c55b498010