Clean up construction of Index's public_targets
Summary:
`Index::new` creates an `Index` called `tmp` containing an empty `public_targets`, then creates a separate nonempty `public_targets`, then performs a [functional update](https://doc.rust-lang.org/1.88.0/reference/expressions/struct-expr.html#functional-update-syntax) to create a second `Index` containing the new nonempty `public_targets` combined with the rest of the fields from `tmp`.

https://www.internalfb.com/code/fbsource/[cf9be6adcf0093977b571f0360f2c594500c1476]/fbcode/common/rust/tools/reindeer/src/index.rs?lines=60%2C64%2C89%2C97-98%2C114-116%2C129%2C141-145%2C309

That arrangement is a bad fit for the logic that is going to be added to `Index::new` in {D77062707} and the rest of this stack. It is better to deal with a single `Index` and progressively add information to it during the course of `Index::new`.

Reviewed By: diliop

Differential Revision: D77062711

fbshipit-source-id: f9c90753ffd0e75469c6cdf34639342710135d0b