Add a resolver from ManifestDep to NodeDep
Summary:
In order to perform per-platform feature resolution in Reindeer in a future diff, we need the ability to associate a ManifestDep (edge in the unresolved dependency graph, i.e. Cargo.toml [dependencies] sections) with the corresponding NodeDep (edge in the resolved dependency graph), which is not straightforward.

This is needed because only the unresolved dependency graph contains information about what features are implied by a dependency **edge**, such as the following one.

```lang=toml
[dependencies]
example = { version = "1", features = ["feat"] }
```

In contrast, the resolved dependency graph contains feature information associated with **nodes** only, i.e. the final set of features chosen by Cargo for each package in the graph, which Reindeer currently uses but should not.

Reviewed By: diliop

Differential Revision: D77062705

fbshipit-source-id: dbb8c18b6e29b48ae17716b9a02ee9788eeca48c