Perform custom feature resolution per platform
Summary:
This logic is in charge of figuring out, for each platform known to Reindeer, what features are enabled for each crate present in the dependency graph on that platform.

For example consider a set of packages with "a" being a root workspace member:

```lang=toml
name = "a"

[dependencies]
c = "*"

[target.'cfg(target_os = "example")'.dependencies]
b = "*"
```

```lang=toml
name = "b"

[dependencies]
c = { version = "*", features = ["feat"] }
```

```lang=toml
name = "c"

[features]
feat = []
```

Reindeer would need to determine that c's "feat" feature is enabled on all platforms for which `target_os = "example"`, and not enabled on all other platforms.

Reviewed By: diliop

Differential Revision: D77062707

fbshipit-source-id: 918bc2bd1e28dc4eb28c57d5f7af4dc74d571198