Distinguish platforms that are execution platforms
Summary:
When processing a crate like this:

```lang=log
name = "a"

[dependencies]
b = { version = "*", default-features = false }

[build-dependencies]
b = { version = "*", features = ["std"] }
```

it is necessary for Reindeer to know the relationship between target platforms and execution platforms. Suppose there are 2 platforms, a no-std embedded platform and a Linux execution platform. When configuring "a" for the embedded platform, it is necessary to know to enable "b"'s std feature on the Linux platform which will be used for the build script, and not on the target platform. Similar for proc macro dependencies.

This diff makes every platform configurable in reindeer.toml with a boolean saying whether it is an execution platform.

```lang=toml
[platform.linux-x86_64]
execution-platform = true

[platform.thumbv7em-none-eabi]
execution-platform = false
```

Every platform that is an execution platform will be considered a potential execution platform for every platform known to Reindeer.

I considered a more elaborate configuration where each individual platform could say which set of other platforms serve as its execution platforms:

```lang=toml
[platform.linux-x86_64]
execution-platforms = ["linux-x86_64"]

[platform.thumbv7em-none-eabi]
execution-platforms = ["linux-x86_64"]
```

or vice versa, which platforms each platform is the execution platform for:

```lang=toml
[platform.linux-x86_64]
execution-platform-for = ["linux-x86_64", "thumbv7em-none-eabi"]

[platform.thumbv7em-none-eabi]
```

but the simpler configuration is good enough and easier to reason about. For example it can be easily determined that we are not running rustc on any thumbv7em-none-eabi host.

Reviewed By: diliop

Differential Revision: D77685015

fbshipit-source-id: e548c22b06f68df7bdbfad77a2a08eeb1db9a12e