Set default feature in universe unless specified otherwise
Summary:
There was a weird coupling between the universe's name and the set of features. A universe named "DEFAULT" would always be configured with the `default` feature (and no way to say otherwise). Any universe not named "DEFAULT" would get `default` disabled by default, unless enabled by specifying `features = ["default"]` in the universe.

To me it makes more sense if universe names are arbitrary strings without any significance to Reindeer. The features for a universe should be controlled through the universe's `features` only, not its name.

This diff gives you the `default` feature if no `features = [...]` is specified for the universe, otherwise it gives you the features that the universe has specified.

Reviewed By: cjlongoria

Differential Revision: D77880299

fbshipit-source-id: f8fe439d1bf330eec17299bb6e9de1c494cfc94b