Add a way to set features by platform
Summary:
As part of removing Reindeer universes in favor of platform-specific feature resolution, we need a way for platforms to specify what features to enable when buckifying that platform. For example our default universe currently enables "default" and our no_std universe enables "universe-no_std". This will instead be associated with a platform. Linux-platform010 will enable "default" and x86_64-unknown-none will enable "universe-no_std".

This diff adds a way to pin a platform to one specific universe, and that chosen universe's `features` and `include_crates` will always be used on this platform, regardless of universe.

```lang=toml
[platform.x86_64-unknown-none]
execution-platform = false
universe = "no_std"
```

This is a temporary workaround that is the best way to rapidly remove a bunch of universe-related logic. Coming up in about 20 more diffs, universes will be gone altogether, `include_crates` will be gone in favor of platform-specific `[target.cfg(…).dependencies]` dependencies, and this `platform.*.universe` reindeer.toml option will be replaced with just `platform.*.features` which are the features to enable on a particular platform.

Reviewed By: diliop

Differential Revision: D77904052

fbshipit-source-id: 59d480f44289603afe63b82bfbaa110a60bee1d1