Do not store ManifestTarget inside Fixups
Summary:
A fixups.toml file is not specific to any single `ManifestTarget` within a package (for example its library crate, or a specific binary crate). The same fixups.toml applies to every target in the package.

Previously `Fixups` was being constructed holding a particular `ManifestTarget` because all the code in buckify.rs operates in terms of targets.

But the feature resolver only operates in terms of packages, because features are a property of a package, not a target. In order for the feature resolver to look at fixups in {D78124833}, this diff removes the `ManifestTarget` from `Fixups` and instead passes it to the methods which consider a specific target.

Reviewed By: diliop

Differential Revision: D78124840

fbshipit-source-id: fdb5262b5d37f65bf3d32431ab6775331e44e4ec