Cache fixups.toml file loading
Summary: We are re-parsing the same fixups.toml files over and over. It does not significantly affect performance because the vast majority of Reindeer runtime is spent inside `cargo metadata` anyway, not in Reindeer code. But the real reason I need this refactor is for {D78222867}. Keeping track of which globs in which fixups.toml files have ever produced a match is a lot easier if each fixups.toml file is only loaded once and we can use an `AtomicBool` stored inside the glob to keep track of matching.

Reviewed By: diliop

Differential Revision: D78222866

fbshipit-source-id: 0f94e3ae0bb42b1e02b636473f3396b65f75e771