Defer diagnosis of unused globs
Summary:
As part of doing the refactor from:

```lang=toml
['cfg(all(all(all())))']
version = ">=0.6"
...
```

to:

```lang=toml
['cfg(version = ">=0.6")']
...
```

one of the consequences is that every conditional fixup needs to be evaluated against every version of the package. We can no longer know ahead of time (before evaluating the cfg) that a conditional fixup will definitely not apply to a particular version.

This requires a new approach for implementing `strict_globs`. Instead of checking eagerly as each glob is walked, we will instead store an AtomicBool inside every glob to remember whether it has matched any files, and we check those all at once at the end of buckify.

This approach works better for many fixups fields that were not previously covered by `strict_globs`. In fbsource/third-party/rust it surfaces 16 unused globs ({D78222868}) that were not previously recognized as unused.

Reviewed By: diliop

Differential Revision: D78222867

fbshipit-source-id: db2eba3e27dbc921ec01d5803da6a5b660d7b6f7