Add method on Index to get compatible platforms for a Manifest
Summary:
Previously, we were relying on `compute_features` to return `None` to indicate that a crate was not configured for the requested platform. When buckifying features and deps for such a crate, we would know to skip that platform.

In the rest of this stack, it helps to have the platform compatibility information more readily available as `BTreeSet<&PlatformName>`, the name of every platform that the crate is compatible with.

Reviewed By: diliop

Differential Revision: D78248602

fbshipit-source-id: 07f9cba421618323cd44a96ca1336fde44b1952a