Precise platform-specific sources computation
Summary:
Reindeer is generating a redundant representation of sources for crates that have a platform-specific extra_srcs fixup.

For example:

https://www.internalfb.com/code/fbsource/[7bd6dec93b45b3f6cc22d69bf83b0697eb8a32bd]/third-party/rust/fixups/termwiz/fixups.toml?lines=1-4

https://www.internalfb.com/code/fbsource/[ca3931effa430d6c8f9fdd9a6f84b8225c8952a4]/third-party/rust/BUCK?lines=223554-223573%2C223575%2C223577-223578%2C223643%2C223668-223669%2C223678-223680%2C223689%2C223708-223709%2C223718-223720%2C223729-223731%2C223740-223742%2C223751-223753%2C223762-223764%2C223773%2C223786-223787%2C223796-223798%2C223807-223809%2C223818-223820%2C223829-223831%2C223840%2C223847-223848%2C223857%2C223888-223889%2C223894%2C223901%2C223937

This diff converts `srcs` to the approach used for `features` by D77062712, and for `deps` by D77062716. This recognizes when the same src is common to every platform that a crate is compatible with, and will avoid using `platforms` for such sources.

Reviewed By: diliop

Differential Revision: D78248582

fbshipit-source-id: d48cd269630fb1fcbd0a948df84c6f9f88fadfb5