Un-sort rustc flags
Summary:
It does not make sense for rustc flags to be represented using `BTreeSet<String>`. Order is important.

For example this fixup:

https://www.internalfb.com/code/fbsource/[600e9a8c38c19d17729a292ae83f79295f3199bc]/xplat/rust/toolchain/sysroot/1.87.0/fixups/panic_abort/fixups.toml?lines=2

happens to render as `rustc_flags = ["-C", "panic=abort"]` because in ASCII order `-C` is less than `panic=abort`. But if someone added a second `-C` flag to it, it would render as `rustc_flags = ["-C", "-C", "panic=abort", "..."]` which is nonsensical.

This diff makes Reindeer preserve the order of rustc flags fixups.

Reviewed By: diliop

Differential Revision: D78248599

fbshipit-source-id: 4030a39ad55f05311952e05007e358e9f99ee58e