Move reindeer platform select to PACKAGE
Summary:
In {D78425948}, more macros need to be able to enumerate all Reindeer platforms used by the current package.

Instead of asking the user of Reindeer to supply their own macros that wrap `buildscript_run` and `buildscript_build` to pass in the set of platform names matching their reindeer.toml, this diff moves platforms to being defined through a PACKAGE / BUCK_TREE file instead. Call `set_reindeer_platforms` to store the `select` that was previously supplied as the third argument to `apply_platform_attrs`.

A workaround is necessary because `write_package_value` does not allow `select`. An RFC for supporting this is discussed in <https://fb.workplace.com/groups/buck2dev/posts/3985827831705259>. For now we just translate the `select` passed by the caller into a `dict` and store that, and translate it back to `select` in `get_reindeer_platforms`.

Reviewed By: diliop

Differential Revision: D78425947

fbshipit-source-id: 67ad67f78b9ee315e6c1b9c070b3112faec1b708