Add `--fast` codepath based on cargo-as-a-library
Summary:
The vast majority of the time taken by `reindeer buckify` (more than 90%) is spent waiting for `cargo metadata`.

Cargo's implementation of `cargo metadata` has pathological behavior in workspaces that have all three of the following:

1. Many distinct package sources (**S**). This is usually the number of distinct git repositories in the dependency graph, plus 1 for the crates.io source.

2. Many packages (**P**) in the transitive dependency graph.

3. A `.cargo/config.toml` that performs a many-to-one remapping of sources to a local directory source, typically a `vendor` directory.

In this scenario, the runtime of `cargo metadata` is ***O*(S × P)** instead of the expected ***O*(S + P)**.

Unfortunately fbsource/third-party/rust is affected. In our use case, currently S is 62 and P is 3662, and we are vendoring. As such, Cargo commands in this workspace end up being 62× slower than they should be.

https://www.internalfb.com/code/fbsource/[41ec9a269c699c9788fc4b900144046d781f3c48]/third-party/rust/.cargo/config.toml?lines=14-17%2C309-310

This diff adds a `--fast` codepath that reimplements most of the logic of `cargo metadata` using the APIs exposed by the `cargo` library crate.

For reasons that I have not tracked down, this diff unexpectedly *already* speeds up `reindeer buckify` in fbsource/third-party/rust. The old codepath based on `cargo metadata` takes about 100 seconds while the `--fast` codepath takes 75 seconds. This was not intentional, since the diff is only a direct reimplementation of the existing `cargo metadata` behavior for the purpose of optimizing it in other diffs in the rest of this stack.

Reviewed By: cfsmp3

Differential Revision: D78619273

fbshipit-source-id: b67a1863b488f59dbe2374910d03c653813b890a