Share packages between every source that has the same replacement
Summary:
This optimization reduces the time for `reindeer buckify` in fbsource/third-party/rust to 17 seconds.

Of this, about 4.5 seconds are spent loading the lockfile and manifests (previously 88 seconds for the old codepath), 4.5 seconds are spent doing feature resolution for 32 platforms, and 8 seconds are spent in generating rules (mostly in srcfiles).

Both the feature resolver and rule generator have never been optimized, since the runtime for buckify has always been dominated by `cargo metadata`. I am optimistic that there is a 10× improvement in each of those, which would bring total `reindeer buckify` time down to 6 seconds.

Reviewed By: cfsmp3

Differential Revision: D78623154

fbshipit-source-id: 1afe28462530544f541971b3cca3540a2bf81c59