Move fixup_dir from Fixups into FixupConfigFile
Summary:
Every time anything looked at a fixup (`FixupsCache::get`), this code would spend time pasting together path segments to compute the `fixup_dir` filepath. This happens 568185 times in buckifying fbsource/third-party/rust, adding up to over 0.5 seconds of runtime, which at this point is 4% of buckify's entire runtime.

It is a simple change to move `fixup_dir` inside of `FixupConfigFile` so that it is only computed the first time that fixup is loaded, and can be used as a reference in all subsequent uses of the same fixup.

Reviewed By: cfsmp3

Differential Revision: D78636033

fbshipit-source-id: c3d4c2b28639a40bc17acaf49fa7c148a7d5fc68