Support crate_root that is not inside manifest directory
Summary:
I need this for the `rustc_proc_macro` crate in Rust 1.89.0.

https://www.internalfb.com/code/fbsource/[6e345ae29c9170b2f5968d03bbef6cb0f60eec30]/third-party/tp2/rust/1.89.0%2Bllvm.19/rustc-1.89.0-src/compiler/rustc_proc_macro/Cargo.toml?lines=7-8%2C12-13

Without this fix, `reindeer buckify` generates:

```lang=starlark
rust_library(
    name = "rustc_proc_macro-0.0.0",
    ...
    crate_root = "compiler/rustc_proc_macro/../../library/proc_macro/src/lib.rs",
    ...
)
```

which does not work because there are no source files located under "compiler/rustc_proc_macro", so that directory is not going to exist in the remapped srcs directory fed to rustc and "compiler/rustc_proc_macro/.." cannot be resolved.

The correct Buck target for this crate would be:

```lang=starlark
rust_library(
    name = "rustc_proc_macro-0.0.0",
    ...
    crate_root = "library/proc_macro/src/lib.rs",
    ...
)
```

Reviewed By: diliop, capickett

Differential Revision: D80627283

fbshipit-source-id: 7fd7f19c59c8737d492293dc7b03c4141de350df