drop trailing slash in `mapped_manifest_dir` for git sources
in git sources, when the repo root is the same as the manifest dir,
currently reindeer with attempt to join the repo name with an empty path
(the manifest sub-path), which produces a path with a trailing slash,
due to how `Path::join` works.

This path is later used to create the key in a buck2 file-to-source
map-dict, which fails, since buck2 does not want trailing slashes in the
keys for those maps.

this detects if the manifest sub-path would be empty, and skips joining
it.