Recognize rustc-link-lib and rustc-link-search
Summary:
Requires an opt-in in fixup.toml, because this is usually not the behavior you would want when using buckified dependencies.

```lang=toml
[buildscript.run]
# Whether to translate cargo::rustc-link-lib into `-l` flags for rustc
rustc_link_lib = true
# Whether to translate cargo::rustc-link-search into `-L` flags for rustc
rustc_link_search = true
```

This is the set of crates in fbsource/third-party/rust for which this would make a difference if enabled:

- `getrandom 0.1.15`

```lang=text
-lframework=Security
```

- `gmp-mpfr-sys 1.6.4`

```lang=text
-Lnative=$(abspath buck-out/v2/gen/fbsource/f5cab21ee3fa053c/third-party/rust/__gmp-mpfr-sys-1.6.4-build-script-run__/OUT_DIR/lib)
-lstatic=gmp
```

- `riscv-rt 0.15.0`

```lang=text
-L$(abspath buck-out/v2/gen/fbsource/74dcf35af8365fbc/third-party/rust/__riscv-rt-0.15.0-build-script-run__/OUT_DIR)
```

- `wit-bindgen-rt 0.39.0`

```lang=text
-lwit_bindgen_cabi_realloc0_39_0
-Lnative=$(abspath buck-out/v2/gen/fbsource/8340c3c93efad46f/third-party/rust/__wit-bindgen-rt-0.39.0-build-script-run__/OUT_DIR)
```

In fbsource, we do not want these flags. Instead the linking is usually handled via `extra_deps` pointing to the required dependency built through Buck.

https://www.internalfb.com/code/fbsource/[cab8e2cfe3514d6118933087db5cd911aeaac1f5]/third-party/rust/fixups/getrandom/fixups.toml?lines=10-13

https://www.internalfb.com/code/fbsource/[cab8e2cfe3514d6118933087db5cd911aeaac1f5]/third-party/rust/fixups/gmp-mpfr-sys/fixups.toml?lines=1-3

In the case of `-lgmp` this is C library built from source.

https://www.internalfb.com/code/fbsource/[9eb65901fef072046247e689531c9430f0b5fb90]/third-party/gmp-6.3.0/BUCK?lines=71-73%2C85%2C96%2C123

In the case of `-lframework=Security` it is a set of `exported_linker_flags` packaged in a `prebuilt_cxx_library`.

https://www.internalfb.com/code/fbsource/[9eb65901fef072046247e689531c9430f0b5fb90]/third-party/toolchains/macos/BUCK?lines=115-117

https://www.internalfb.com/code/fbsource/[9eb65901fef072046247e689531c9430f0b5fb90]/third-party/toolchains/macos/defs.bzl?lines=17-19%2C33-37%2C39%2C41-42

Reviewed By: diliop

Differential Revision: D83386224

fbshipit-source-id: 1ef863777e1d8a2c8b117253396609e035a22d9c