Fix overcounting of extra_deps that duplicate an existing dep
Summary:
If some package had a fixups.toml containing an `extra_deps` that was already one of the package's dependencies according to Cargo (for example `extra_deps = [":indexmap-2.11.4"]` for a package that already had `indexmap` in its dependencies) then that dependency was previously being duplicated in the collection returned by `compute_deps`:

https://www.internalfb.com/code/fbsource/[8f491cb88dc1a2e30591b8eff161f7349f4cc377]/fbcode/common/rust/tools/reindeer/src/fixups.rs?lines=747%2C752-757

The one from the fixup was inserted here:

https://www.internalfb.com/code/fbsource/[8f491cb88dc1a2e30591b8eff161f7349f4cc377]/fbcode/common/rust/tools/reindeer/src/fixups.rs?lines=778-781%2C784-785

The one from the Cargo deps was inserted here:

https://www.internalfb.com/code/fbsource/[8f491cb88dc1a2e30591b8eff161f7349f4cc377]/fbcode/common/rust/tools/reindeer/src/fixups.rs?lines=760-761%2C837%2C841-842%2C853-855%2C862-863

As a consequence, the duplicated dependency would be double-counted in `dep_in_how_many_platforms`:

https://www.internalfb.com/code/fbsource/[531cb6928a90a802af08d59bfa8068296cac2e38]/fbcode/common/rust/tools/reindeer/src/buckify.rs?lines=566-567%2C581-583

and then Reindeer would incorrectly end up treating it as a platform-specific dependency when it wasn't, or vice versa. For example if we have 28 Reindeer platforms, `dep_in_how_many_platforms[&key]` for a duplicated dependency could end up being 56. Inversely, for a platform-specific dependency that was supposed to apply to exactly half of the platforms (14), it could end up with `dep_in_how_many_platforms[&key]` being 28 making it look platform-independent.

https://www.internalfb.com/code/fbsource/[531cb6928a90a802af08d59bfa8068296cac2e38]/fbcode/common/rust/tools/reindeer/src/buckify.rs?lines=591-597

This diff deduplicates the collection items returned by `compute_deps` using the same key that is used by `dep_in_how_many_platforms`.

Reviewed By: cjlongoria

Differential Revision: D84363314

fbshipit-source-id: 6b52f9a34efcd7805946f193d447a2e9a1da696e